moved manager init to controller package
